requirements.txt
streamlit
google-generativeai
pdf2image
pandas
pydantic
python-dotenv
gdown
